+++ front/src/components/Main/LifeQuotes/LifeQuotes.module.scss
