+++ front/src/components/Main/LifeQuotes/LifeQuotes.tsx
